+++ routes/homeStack.js
@@ -1,0 +1,25 @@
+// import React from 'react';
+// import { createAppContainer } from 'react-navigation';
+// import { createStackNavigator } from 'react-navigation-stack';
+// import About from '../screens/about';
+
+// const screens = {
+//     About: {
+//         screen: About
+//     }
+// }
+
+// const HomeStack = createStackNavigator(screens);
+
+// <NavigationContainer>
+//     <Stack.Navigator>
+//         <Stack.Screen
+//             name="About"
+//             component={aboutScreen}
+//             options={{ title: 'About This App' }}
+//         />
+//     </Stack.Navigator>
+// </NavigationContainer>
+
+// export default createAppContainer(HomeStack);
+
